requirements.txt
azure-storage-blob==12.6.0
Flask==1.1.2
numpy==1.18.2
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
python-dotenv==0.15.0
Werkzeug==1.0.1